🔧 build(c): clear -Wall/-Wextra warnings and gate them in CI
A plain `uv build` emitted a stream of compiler warnings, the loudest being a
`*line/*col` fragment in a tree.h comment whose `/*` reads as a nested block
comment and re-fires for every translation unit that includes the header. The
rest were unused `handle`/`tree` bindings the per-tree critical section only
consumes on free-threaded builds, and a genuine bug: buf_ensure assigned the
result of PyMem_Resize back into `grown`, but the macro already assigns to its
argument, so `grown` took two unsequenced writes.

Reword the comments, drop or void-cast the lock-only bindings, remove the dead
`tree` parameter, let the macro own the assignment, and mark the XPath axis
fall-through with a TH_FALLTHROUGH attribute so gcc's -Wimplicit-fallthrough
stops flagging the prose comment. Widen each operand before the add in the SSE2
block_extra so clang-tidy's misplaced-widening-cast check is satisfied.

Nothing enforced this, so warnings accrued silently. clang-tidy ran with `-*`
as its base, which also disabled the clang-diagnostic family, and the hook is
skipped on pre-commit.ci. Add a `warnings` tox env that compiles with
`-Dwerror=true` (level 2 only, so the API-mandated void*/function-pointer slot
casts under -Wpedantic stay out of scope) and run it across a compiler x
SIMD-path grid in CI: gcc and clang over x86/SSE2 and arm64/NEON, plus MSVC.
Enable clang-diagnostic-* in the tidy config and run that hook on both x86 and
arm64 runners, since the #if-gated SIMD helpers are only analyzed on the arch
that builds them.

+++ src/turbohtml/_c/serialize/escape.c
@@ -121,7 +121,7 @@ static inline Py_ssize_t block_extra(const uint8_t *block, int quote) {
         extras = _mm_add_epi8(extras, _mm_and_si128(_mm_cmpeq_epi8(bytes, _mm_set1_epi8('\'')), _mm_set1_epi8(5)));
     }
     __m128i sums = _mm_sad_epu8(extras, _mm_setzero_si128());
-    return (Py_ssize_t)(_mm_cvtsi128_si32(sums) + _mm_extract_epi16(sums, 4));
+    return (Py_ssize_t)_mm_cvtsi128_si32(sums) + (Py_ssize_t)_mm_extract_epi16(sums, 4);
 }
 
 /* One mask bit per byte; SPECIAL_INDEX / SPECIAL_CLEAR hide the layout. */
